Update README.md
This Digital Clock is a Graphical user interface(GUI) application that shows current date and time to the user using labels. The clock updates the date and time automatically. The time is shown in HH:MM:SS format i.e., hours, minutes and seconds respectively. The date is in the format of EEE, MMM dd, yyyy i.e., day, month, day and year respectively.

The Java Swing library which gives a flexible and powerful set of components is used to develop the Digital Clock. The application makes use of Swing components such as JLabel and JPanel among many other Swing components. Date and time is set to the JLabel using setText() Method. Methods of Calendar class in util package is used to get the current date and time.  

Methods used in the project:

getTime():

This is the method of Calendar class which belongs the java.util package. It returns the current system date and time in the given format.